% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{complexity_tree_fun}
\alias{complexity_tree_fun}
\title{Find complexity of a dtree object}
\usage{
complexity_tree_fun(object.dtree)
}
\arguments{
\item{object.dtree}{Object of class dtree}
}
\value{
Numeric value that is the complexity (number of branchnodes)
}
\description{
Find complexity of a dtree object
}
