% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{cut_tree_fun}
\alias{cut_tree_fun}
\title{Reduce depth of dtree if possible}
\usage{
cut_tree_fun(object.dtree)
}
\arguments{
\item{object.dtree}{Object of class dtree}
}
\value{
Object of class dtree
}
\description{
If there is a whole depth layer of 'branchnodes' that contains no branch node, this dimension is removed
}
