% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dtree_class.R
\docType{class}
\name{dtree-class}
\alias{dtree-class}
\title{An S4 class to represent a decision tree.}
\description{
An S4 class to represent a decision tree.
}
\section{Slots}{

\describe{
\item{\code{A}}{A numeric matrix with p rows and q columns, where each column represents a branch node.
Each column contains at most one '1', which corresponds to the feature on which the split is made.}

\item{\code{bvec}}{A length q numeric vector where each entry represent a branch node and the split value
(if the split is on a numeric feature).}

\item{\code{cats}}{A list of length q, where each list entry represents a branch and is a character array
(if the split is on a categorical feature).}

\item{\code{dep_var}}{A character string for the dependent variable}

\item{\code{leaf_classes}}{The predicted class for point in certain leaf node}

\item{\code{features}}{A character vector of all p feature on which the tree is built.}
}}

