% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search_functions.R
\name{find_best_parallel_split}
\alias{find_best_parallel_split}
\title{Find best parallel split}
\usage{
find_best_parallel_split(object.subdtree, datasub, weightssub, minleafsize,
  misclassification_weights)
}
\arguments{
\item{object.subdtree}{Object of class dtree. The split of the rootnode will be optimized.}

\item{datasub}{A dataframe: all observations that reach the rootnode of the subtree}

\item{minleafsize}{The minimum number of observations in any terminal <leaf> node.}
}
\value{
A list of two, where the first element is an object of class dtree and the second
element is a logical value indicating if the subtree has changed.
}
\description{
For a specific branchnode (root of inputted tree), find the optimal split.
}
