% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{find_daughter_nodes}
\alias{find_daughter_nodes}
\title{Find daughter nodes}
\usage{
find_daughter_nodes(node)
}
\arguments{
\item{node}{Numeric value}
}
\value{
A numeric array of 2 indicating the direct daughters of node
}
\description{
Find daughter nodes
}
