% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{find_parent_node}
\alias{find_parent_node}
\title{Find parent node}
\usage{
find_parent_node(node)
}
\arguments{
\item{node}{Numeric value}
}
\value{
A numeric value indicating the direct parent of node
}
\description{
Find parent node
}
