% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{find_parent_nodes}
\alias{find_parent_nodes}
\title{Find parent nodes}
\usage{
find_parent_nodes(node)
}
\arguments{
\item{node}{Numeric value}
}
\value{
A numeric array of all parent of node
}
\description{
Find parent nodes
}
