% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{find_pos_points}
\alias{find_pos_points}
\title{Find position of all data points in tree}
\usage{
find_pos_points(object.dtree, data)
}
\arguments{
\item{object.dtree}{Object of class dtree}

\item{data}{A dataframe with features and dependent variable class.}
}
\value{
A matrix in which each row indicated a observation and each column indicates a node. A 1 in
position (i, j) indicated that observation i is in node j.
}
\description{
Find position of all data points in tree
}
