% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{find_subtree}
\alias{find_subtree}
\title{Find subtree rooted at node}
\usage{
find_subtree(rootnode, object.dtree)
}
\arguments{
\item{rootnode}{Rootnode of subtree}

\item{object.dtree}{Object of class dtree}
}
\value{
A list with object.subdtree being an object of class dtree and the subtree rooted at the rootnode
inputted, branchnodes is the a numeric array with all branchnodes of the subtree, and leafnode is a logical
indicating wheter the rootnode is a leaf node
}
\description{
Find subtree rooted at node
}
