% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{find_where_and_leafs}
\alias{find_where_and_leafs}
\title{Basic analysis of performance of dtree ob dataset}
\usage{
find_where_and_leafs(object.dtree, data, weights,
  misclassification_weights)
}
\arguments{
\item{object.dtree}{Object of class dtree}

\item{data}{A dataframe with features and dependent variable class.}
}
\value{
A list with leaf_classes an factor array of classes prescribed by each leaf node,
misclass a numeric array of the misclassification on the data associated with each leaf node, and
minleafsize the minimum leaf size in a leaf node.
}
\description{
Finds the classes associated with leaf nodes, the misclassification per leaf node and the minimum
leaf size.
}
