% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{get_class_node}
\alias{get_class_node}
\title{Find the classification from a set of observations}
\usage{
get_class_node(obs)
}
\arguments{
\item{obs}{A numeric array with classes}
}
\value{
The most prevalent class
}
\description{
Find the classification from a set of observations
}
