% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search_functions.R
\name{local_search}
\alias{local_search}
\title{Perform local search procedure to find OCT}
\usage{
local_search(object.dtree, data, weights = NULL, alpha = 0,
  minleafsize = 1, maxdepth = 6, misclassification_weights = NULL)
}
\arguments{
\item{object.dtree}{Object of class dtree, usually a tree created by CART procedure
where each split only a random subset of features is considered.}

\item{data}{A dataframe: all character variables must be factors, and the dependent
variable must be a factor.}

\item{alpha}{A complexity parameter}

\item{minleafsize}{The minimum number of observations in any terminal <leaf> node.}

\item{maxdepth}{Set the maximum depth of any node of the final tree,
with the root node counted as depth 0.}
}
\value{
A list of two, where the first element is an object of class dtree and the second
element is the loss associated with the dtree.
}
\description{
The local search procedure designed by Dunn takes a CART decision tree as input, and then exhaustively
considers changes to each node to optimize the tree.
}
