% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CART_functions.R
\name{makeCART}
\alias{makeCART}
\title{CART procedure - random subset of features considered}
\usage{
makeCART(data, minleafsize, maxdepth, features = NULL, dep_var = NULL)
}
\arguments{
\item{data}{A dataframe: all character variables must be factors, and the dependent
variable must be a factor.}

\item{minleafsize}{The minimum number of observations in any terminal <leaf> node.}

\item{maxdepth}{Set the maximum depth of any node of the final tree,
with the root node counted as depth 0.}

\item{features}{The features that should be considered. If empty the first n - 1 column
names of the dataframe will be taken as features.}

\item{dep_var}{The dependent variable. If empty the last column
name of the dataframe will be taken as dependent variable.}
}
\value{
An object of class dtree
}
\description{
CART procedure - random subset of features considered
}
