% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalization.R
\name{optimaltrees}
\alias{optimaltrees}
\title{Create the optimal classification tree (OCT)}
\usage{
optimaltrees(formula, data, weights = NULL, maxdepth = 6,
  minleafsize = 1, numbertries = 10,
  misclassification_weights = NULL)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class):
a symbolic description of the model to be fitted.}

\item{data}{A dataframe: all character variables must be factors, and the dependent
variable must be a factor. NA rows will be deleted.}

\item{maxdepth}{Set the maximum depth of any node of the final tree,
with the root node counted as depth 0.}

\item{minleafsize}{The minimum number of observations in any terminal <leaf> node.}

\item{numbertries}{The number of searches that will be performed at each possible depth,
to tune the parameters and find the optimal tree.}
}
\value{
An object of class dtree that is the optimal classification tree
}
\description{
This function created an interpretable decision tree to classify data. It created batches of trees of specific
depths to optimize the parameters alpha (complexity parameter) and maximum depth of the tree, then creates
a new batch of trees with these optimal parameters. The best performing tree from this batch is returned.
}
