% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search_functions.R
\name{optimize_node_parallel}
\alias{optimize_node_parallel}
\title{Perform Optimize Node Parallel procedure}
\usage{
optimize_node_parallel(subtree, datasub, weightssub, data, weights, alpha,
  minleafsize, misclassification_weights)
}
\arguments{
\item{subtree}{A list with an object of class dtree, the associated branchnodes
(result of function find_subtree) and a logical indicating if the subtree is a leaf node.}

\item{datasub}{A dataframe: all observations that reach the rootnode of the subtree}

\item{data}{A dataframe: the full dataset}

\item{alpha}{A complexity parameter}

\item{minleafsize}{The minimum number of observations in any terminal <leaf> node.}
}
\value{
A list of two, where the first element is an object of class dtree and the second
element is a logical value indicating if the subtree has changed.
}
\description{
The Optimize Node Parallel procedure checks to see if a change in the split (or leaf) at the rootnode
of a subtree is necessary to reduce the loss associated to the subtree. If the rootnode is a branchnode,
it is checked whether 1) there exists a better split, 2) it is better to replace with the lower subtree, or
3) it is better to replace with upper subtree. If the rootnode is a leaf, it is checked wheter it is better
to apply a split.
}
