% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare dataset for local search procedure}
\usage{
prepare_dataset(data, weights, features, dep_var)
}
\arguments{
\item{data}{A dataframe}

\item{features}{Character array of the features that will be used in training the OCT}

\item{dep_var}{Character that indicated the dependent variable (class name)}
}
\value{
A list with data being a dataframe with normalized numeric features, and all other variables factors,
and a matrix with the minimum and maximum values of all numeric features.
}
\description{
This function normalized all numeric features from 0 to 1 and forces all character features to be factors,
as well as forcing the dependent variable (class) to be a factor
}
