% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalization.R
\name{prune}
\alias{prune}
\title{Prune OCT}
\usage{
prune(object.dtree, trainingdata, validationdata, trainingweights,
  validationweights, misclassification_weights)
}
\arguments{
\item{object.dtree}{An object of class dtree}

\item{trainingdata}{Dataframe that will be used to train the OCT}

\item{validationdata}{Dataframe that will be used to validate the OCT}
}
\value{
A list of two: vector alpha of critical values for complexity parameter and vector of corresponding
validation errors for each critical value
}
\description{
For each tree, there is a corresponding interval for alpha from which any value will generate this tree.
This procedure finds the discrete values of alpha for which the tree changes, and for each range of alpha,
finds the validation error.
}
