% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dtree_class.R
\name{visualize_tree}
\alias{visualize_tree}
\title{Visualize decision tree}
\usage{
visualize_tree(object.dtree)
}
\arguments{
\item{object.tree}{an object of class \code{dtree}}
}
\description{
\code{visualize_tree} plots the decision tree including splits and
leaf assignments in the plot window and prints splits and leafs in command window.
}
\examples{
A <- matrix(0, nrow = 4, ncol = 3); A[3,1] <- 1; A[4,3] <- 1
tree1 <- new("dtree", A = A, bvec = c(2.5, 0, 1.8), cats = vector('list', 3),
             dep_var = 'Species', leaf_classes = c(NA, 'setosa', 'versicolor', 'virginica'),
             features = c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width"))
visualize_tree(tree1)
}
