% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dtree_class.R
\name{visualize_tree_data}
\alias{visualize_tree_data}
\title{Visualize decision tree with data}
\usage{
visualize_tree_data(object.dtree, data, weights)
}
\arguments{
\item{data}{dataframe}

\item{weights}{weights}

\item{object.tree}{an object of class \code{dtree}}
}
\description{
\code{visualize_tree} plots the decision tree including splits and
leaf assignments in the plot window and prints splits and leafs in command window.
}
