% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dtree_class.R
\docType{methods}
\name{predict,dtree-method}
\alias{predict,dtree-method}
\title{Predict Method for dtree Fits}
\usage{
\S4method{predict}{dtree}(object, newdata, ...)
}
\description{
Predict Method for dtree Fits
}
